pyproject.toml
[tool.poetry]
name = "ai-crew-connect"
version = "0.1.0"
description = ""
authors = ["Nuraly <nuralyhug@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
firebase-admin = "^6.5.0"
python-dotenv = "^1.0.1"
openai = "^1.28.0"
pydantic = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
